requirement.txt
streamlit==1.28.0
langchain==0.0.339
langchain-google-genai==0.0.7
google-generativeai==0.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.26.0
pydantic==2.4.2